# Optimized Stable Diffusion 1.5 example with CLI