--- a/06_gpu/stable_diffusion_cli.py
+++ b/06_gpu/stable_diffusion_cli.py
@@ -29,17 +29,13 @@ app = typer.Typer()
 # into the image. This is technique that allows you to load models much faster
 # by using our [high-performance blob storage file server](https://github.com/modal-labs/blobnet).
 
-model_cache_path = "/model_cache"
-image = modal.Image.debian_slim().run_commands(
+image = modal.Image.debian_slim().apt_install(["curl"]).run_commands(
     [
     "pip install torch --extra-index-url https://download.pytorch.org/whl/cu117",
-    "pip install diffusers[torch] transformers ftfy"]
+    "pip install diffusers[torch] transformers ftfy accelerate"]
 ).run_commands([
-    f"""python -c 'import diffusers; import os; euler = diffusers.EulerAncestralDiscreteScheduler.from_config("runwayml/stable-diffusion-v1-5",subfolder="scheduler",use_auth_token=os.environ["HUGGINGFACE_TOKEN"], cache_dir="{model_cache_path}"); euler.save_config("{model_cache_path}")'""",
-    f"""python -c 'import diffusers; import os; import torch; pipe = diffusers.StableDiffusionPipeline.from_pretrained("runwayml/stable-diffusion-v1-5",use_auth_token=os.environ["HUGGINGFACE_TOKEN"],revision="fp16",torch_dtype=torch.float16,cache_dir="{model_cache_path}"); pipe.save_pretrained("{model_cache_path}")'"""
-], secrets=[modal.Secret.from_name("huggingface-secret")]).run_commands([
-    f"ls -lah {model_cache_path}"
-])
+    "curl -L https://gist.github.com/luiscape/36a8cd29b8ed54cfbfcf56d51fe23cc0/raw/a6bf16996efe7c59114eea7944b0f99741d83d54/download_stable_diffusion_models.py | python"
+], secrets=[modal.Secret.from_name("huggingface-secret")])
 stub.image = image
 
 # ## Global context
@@ -58,24 +54,23 @@ if stub.is_inside():
     torch.backends.cudnn.benchmark = True
     torch.backends.cuda.matmul.allow_tf32 = True
 
-    model_id = "/model_cache"
-    euler = diffusers.EulerAncestralDiscreteScheduler.from_config(
-        model_id,
+    cache_path = "/vol/cache"
+    euler = diffusers.EulerAncestralDiscreteScheduler.from_pretrained(
+        cache_path,
         subfolder="scheduler",
-        cache_dir=model_cache_path)
+        cache_dir=cache_path)
     PIPE = diffusers.StableDiffusionPipeline.from_pretrained(
-        model_id, torch_dtype=torch.float16, scheduler=euler, cache_dir=model_cache_path).to("cuda")
-    PIPE.enable_attention_slicing()
+        cache_path, torch_dtype=torch.float16, scheduler=euler, cache_dir=cache_path).to("cuda")
+    # PIPE.enable_attention_slicing()
 
 
 # This is our Modal function. The function runs through the `StableDiffusionPipeline` pipeline.
 # It sends the PIL image back to our CLI where we save the resulting image in a local file.
 
-@stub.function(gpu=True)
-def _run_inference(prompt:str) -> str:
+@stub.function(gpu=modal.gpu.A100())
+def _run_inference(prompt:str, steps:int = 20) -> str:
     with torch.inference_mode():
-        with torch.autocast("cuda"):
-            image = PIPE(prompt, num_inference_steps=20, guidance_scale=7.0).images[0]
+        image = PIPE(prompt, num_inference_steps=steps, guidance_scale=7.0).images[0]
 
     return image