This adds an optimized Stable Diffusion 1.5 tutorial to our docs. The breakdown is:

- **A100**: 6.5s to load model and 1.6s per generation
- **T4**: 13s to load model and 3.7s per generation

One of the key changes is to use the pre-trained `diffusers.EulerAncestralDiscreteScheduler`. It is fast and generates high quality images with fewer steps (20-30). Other optimizations such as tracing are available [here](https://huggingface.co/docs/diffusers/optimization/fp16#memory-and-speed), so it may be possible to make the generation step faster.